refactor(ViewerContainer): Reorganize logic, convert to TypeScript.
